update levels example
